Implemented Jon's request to use edge weight = 1/#occurrences in SSSP.
git-svn-id: svn+ssh://software-srn.sandia.gov/svn/mapreduce/trunk@321 3fc257bb-8550-0410-81b3-a5c103da83f7

+++ programs/shared.hpp
@@ -71,8 +71,10 @@ public:
   };
 };
 
-// Edge Weight type; set to number of occurrence of edge in original input.
-typedef int WEIGHT;
+// Edge Weight type; set to 
+//   number of occurrence of edge in original input OR
+//   1/number of occurrences of edge.
+typedef float WEIGHT;
 
 // Edge with destination vertex (16-bytes) and edge weight
 // Note:  edge_label1 assumes v is first field of this struct.
